correct deixis
In English we would use "these" in this context.  I can't explain why, but it probably has something to do with the fact that they have not yet been mentioned when the word "these" is spoken/written/heard/read.  "Those" might be used if the terminals had already been mentioned.